Mega Commit due to failed ones
Faster rendering and map from sim save file and long big map sim
Optimised the rendering as much as I could, including recalculating a
load of stuff before the animation begins. Also map saves in sim data
file and is extracted when DisplaySavedSim is called. If it is an old
simulation, you need to set the init mapFile in Test_Graphics_2 to the
name of the mapFile you need for the sim you want to run. Also ran a
1000s sim with big map… took over 2 hours to run though… Also Analyse
needs to be massively optimised.

Merge remote-tracking branch 'origin/master'
# Conflicts: #	Analyse.pyc #	Creature.pyc #	Environment.py #
Environment.pyc #	Test_Graphics_2.pyc

Changing name of TestGraphics to Graphics

Final touches to some parameters to make a nice sim
also fixing push problem hopefully